Latest work on the Loading Chromatograms graph.  It improves existing code by creating bin intervals for comparing peaks, rather than a global comparison that creates a rather sparse graph.  It seems to be faster and animates more smoothly.
Former-commit-id: 9f4f858be4f11ec7f2ae40f3e971653c05e7eae8

+++ pwiz_tools/Skyline/Model/Results/CachedChromatogramDataProvider.cs
@@ -93,9 +93,9 @@ namespace pwiz.Skyline.Model.Results
             }
         }
 
-        public override float? MaxRetentionTime { get { return _maxRetentionTime; } }
+        public override double? MaxRetentionTime { get { return _maxRetentionTime; } }
 
-        public override float? MaxIntensity { get { return _maxIntensity; } }
+        public override double? MaxIntensity { get { return _maxIntensity; } }
 
         public override bool IsProcessedScans
         {
